+++ src/Farmer/Arm/Network.fs
@@ -620,6 +620,7 @@ type NetworkInterface =
 
 
 
+
                             |})
                 |}