+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 ## 1.7.9
 * Container Group: Support for Managed Identity
 * Container App: Support for Managed Identity
+* VMs: Add support for VNets in other resource groups
 
 ## 1.7.8
 * Route Tables: Initial support for Route Tables and Routes